Update generator and refactored some variables
