+++ middleware/test.go
@@ -7,7 +7,6 @@ import (
 	"github.com/gin-gonic/gin"
 )
 
-// Test is just a test middleware.
 func Test() gin.HandlerFunc {
 	return func(c *gin.Context) {
 		if tag, exists := c.GetQuery("tag"); !exists || len(tag) < 1 {
@@ -15,7 +14,7 @@ func Test() gin.HandlerFunc {
 		} else {
 			c.JSON(http.StatusOK, map[string]interface{}{
 				"tag":  tag,
-				"date": time.Now().Format(time.RFC3339[:10]),
+				"date": time.Now().UTC().Format(time.RFC3339[:10]),
 				"uploads": map[string]interface{}{
 					"user_A": 3,
 					"user_B": 2,
